requirements.txt
bs4
flask
flask-cors
gunicorn
google-api-python-client
openai
pandas
python-dotenv
requests
youtube-transcript-api
